debugging delete button still

+++ public/js/delete.js
@@ -1,17 +1,17 @@
 // TO DO: Does not delete expense from response
 const delButtonHandler = async (event) => {
-  // if (event.target.hasAttribute("data-id")) {
-  const id = event.target.getAttribute("data-id");
+  if (event.target.hasAttribute("data-id")) {
+    const id = event.target.getAttribute("data-id");
 
-  const response = await fetch(`/api/expenses/${id}`, {
-    method: "DELETE",
-  });
+    const response = await fetch(`/api/expenses/${id}`, {
+      method: "DELETE",
+    });
 
-  if (response.ok) {
-    location.reload();
-  } else {
-    alert("Failed to delete expense");
-    location.reload();
+    if (response.ok) {
+      location.reload();
+    } else {
+      alert("Failed to delete expense");
+    }
   }
 };
